ssrf: combine subsurface-*-helper into subsurface-helper
combine shared functions like init_ui and run_ui in one file
