+++ subsurface-mobile-helper.cpp
@@ -9,7 +9,6 @@
 #include <QNetworkProxy>
 #include <QLibraryInfo>
 
-#include "core/qt-gui.h"
 
 #include <QQuickWindow>
 #include <QScreen>
@@ -25,8 +24,6 @@
 
 #include "core/ssrf.h"
 
-QObject *qqWindowObject = NULL;
-
 void set_non_bt_addresses() {
 #if defined(Q_OS_ANDROID)
 	connectionListModel.addAddress("FTDI");
@@ -38,101 +35,6 @@ void set_non_bt_addresses() {
 	// this makes debugging so much easier - use the simulator
 	connectionListModel.addAddress("/tmp/ttyS1");
 #endif
-}
-
-void init_ui()
-{
-	init_qt_late();
-}
-
-void run_ui()
-{
-	LOG_STP("run_ui starting");
-	register_qml_types();
-
-	QQmlApplicationEngine engine;
-	LOG_STP("run_ui qml engine started");
-	KirigamiPlugin::getInstance().registerTypes();
-#if defined(__APPLE__) && !defined(Q_OS_IOS)
-	// when running the QML UI on a Mac the deployment of the QML Components seems
-	// to fail and the search path for the components is rather odd - simply the
-	// same directory the executable was started from <bundle>/Contents/MacOS/
-	// To work around this we need to manually copy the components at install time
-	// to Contents/Frameworks/qml and make sure that we add the correct import path
-	QStringList importPathList = engine.importPathList();
-	Q_FOREACH(QString importPath, importPathList) {
-		if (importPath.contains("MacOS"))
-			engine.addImportPath(importPath.replace("MacOS", "Frameworks"));
-	}
-	qDebug() << "QML import path" << engine.importPathList();
-#endif
-	engine.addImportPath("qrc://imports");
-	DiveListModel diveListModel;
-	LOG_STP("run_ui diveListModel started");
-	DiveListSortModel *sortModel = new DiveListSortModel(0);
-	sortModel->setSourceModel(&diveListModel);
-	sortModel->setDynamicSortFilter(true);
-	sortModel->setSortRole(DiveListModel::DiveDateRole);
-	sortModel->sort(0, Qt::DescendingOrder);
-	LOG_STP("run_ui diveListModel sorted");
-	GpsListModel gpsListModel;
-	QSortFilterProxyModel *gpsSortModel = new QSortFilterProxyModel(0);
-	gpsSortModel->setSourceModel(&gpsListModel);
-	gpsSortModel->setDynamicSortFilter(true);
-	gpsSortModel->setSortRole(GpsListModel::GpsWhenRole);
-	gpsSortModel->sort(0, Qt::DescendingOrder);
-	QQmlContext *ctxt = engine.rootContext();
-	ctxt->setContextProperty("diveModel", sortModel);
-	ctxt->setContextProperty("gpsModel", gpsSortModel);
-	ctxt->setContextProperty("vendorList", vendorList);
-	set_non_bt_addresses();
-	LOG_STP("run_ui set_non_bt_adresses");
-
-	ctxt->setContextProperty("connectionListModel", &connectionListModel);
-	ctxt->setContextProperty("logModel", MessageHandlerModel::self());
-
-	engine.load(QUrl(QStringLiteral("qrc:///qml/main.qml")));
-	LOG_STP("run_ui qml loaded");
-	qqWindowObject = engine.rootObjects().value(0);
-	if (!qqWindowObject) {
-		fprintf(stderr, "can't create window object\n");
-		exit(1);
-	}
-	QQuickWindow *qml_window = qobject_cast<QQuickWindow *>(qqWindowObject);
-	qml_window->setIcon(QIcon(":subsurface-mobile-icon"));
-	qqWindowObject->setProperty("messageText", QVariant("Subsurface-mobile startup"));
-	qDebug() << "qqwindow devicePixelRatio" << qml_window->devicePixelRatio() << qml_window->screen()->devicePixelRatio();
-	QScreen *screen = qml_window->screen();
-	QObject::connect(qml_window, &QQuickWindow::screenChanged, QMLManager::instance(), &QMLManager::screenChanged);
-	QMLManager *manager = QMLManager::instance();
-	LOG_STP("run_ui qmlmanager instance started");
-	// now that the log file is initialized...
-	show_computer_list();
-	LOG_STP("run_ui show_computer_list");
-
-	manager->setDevicePixelRatio(qml_window->devicePixelRatio(), qml_window->screen());
-	manager->dlSortModel = sortModel;
-	manager->screenChanged(screen);
-	qDebug() << "qqwindow screen has ldpi/pdpi" << screen->logicalDotsPerInch() << screen->physicalDotsPerInch();
-#if !defined(Q_OS_ANDROID) && !defined(Q_OS_IOS)
-	qml_window->setHeight(1200);
-	qml_window->setWidth(800);
-#endif
-	qml_window->show();
-	LOG_STP("run_ui running exec");
-	qApp->exec();
-}
-
-void exit_ui()
-{
-	delete qApp;
-	free((void *)existing_filename);
-}
-
-double get_screen_dpi()
-{
-	QDesktopWidget *mydesk = qApp->desktop();
-	return mydesk->physicalDpiX();
 }
 
 bool haveFilesOnCommandLine()
